rearrange the file structure and add up
