+++ lib/startup/register.dart
@@ -1,35 +1,36 @@
 import 'package:flutter/material.dart';
-import 'package:freelancer/mainpage.dart';
+import 'package:freelancer/config.dart';
+import 'package:freelancer/startup/login.dart';
 
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 
-class Createjobs extends StatelessWidget {
+class FlRegister extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     // TODO: implement build
     return MaterialApp(
       home: Scaffold(
-        body: Createjobshome(),
+        body: RegisterBody(),
       ),
     );
   }
 }
 
-class Createjobshome extends StatefulWidget {
-  Createjobshome({Key key}) : super(key: key);
+class RegisterBody extends StatefulWidget {
+  RegisterBody({Key key}) : super(key: key);
 
   @override
-  _CreatejobshomeState createState() => _CreatejobshomeState();
+  _RegisterBodyState createState() => _RegisterBodyState();
 }
 
-class _CreatejobshomeState extends State<Createjobshome> {
+class _RegisterBodyState extends State<RegisterBody> {
   String _username;
   String _password;
   String _email;
   String _phone;
   _sendReinfo() async {
-    var apiUrl = "http://10.0.2.2:8080/register?";
+    var apiUrl = "${baseUrl}register?";
     var result = await http.post(apiUrl, body: {
       "username": "$_username",
       "password": "$_password",
@@ -58,7 +59,7 @@ class _CreatejobshomeState extends State<Createjobshome> {
                 IconButton(
                   icon: Icon(Icons.keyboard_arrow_left),
                   onPressed: () {
-                    runApp(FlMainpage());
+                    runApp(FlLogin());
                   },
                 ),
                 Padding(
@@ -67,7 +68,7 @@ class _CreatejobshomeState extends State<Createjobshome> {
                     bottom: 30,
                   ),
                   child: Text(
-                    "发布招聘",
+                    "注册",
                     style: TextStyle(
                       fontSize: 32,
                     ),
@@ -77,8 +78,8 @@ class _CreatejobshomeState extends State<Createjobshome> {
             ),
             TextField(
               decoration: InputDecoration(
-                labelText: "Cate（请输入工种）",
-                prefixIcon: Icon(Icons.category),
+                labelText: "请输入用户名",
+                prefixIcon: Icon(Icons.person),
                 // 未获得焦点下划线设为灰色
                 enabledBorder: UnderlineInputBorder(
                   borderSide: BorderSide(color: Colors.grey),
@@ -94,8 +95,8 @@ class _CreatejobshomeState extends State<Createjobshome> {
             ),
             TextField(
               decoration: InputDecoration(
-                labelText: "请输入标题（简洁明了）",
-                prefixIcon: Icon(Icons.title),
+                labelText: "请输入密码",
+                prefixIcon: Icon(Icons.lock),
                 // 未获得焦点下划线设为灰色
                 enabledBorder: UnderlineInputBorder(
                   borderSide: BorderSide(color: Colors.grey),
@@ -111,8 +112,8 @@ class _CreatejobshomeState extends State<Createjobshome> {
             ),
             TextField(
               decoration: InputDecoration(
-                labelText: "请输入薪资",
-                prefixIcon: Icon(Icons.card_giftcard),
+                labelText: "请输入手机号码",
+                prefixIcon: Icon(Icons.phone),
                 // 未获得焦点下划线设为灰色
                 enabledBorder: UnderlineInputBorder(
                   borderSide: BorderSide(color: Colors.grey),
@@ -128,25 +129,8 @@ class _CreatejobshomeState extends State<Createjobshome> {
             ),
             TextField(
               decoration: InputDecoration(
-                labelText: "请输入位置",
-                prefixIcon: Icon(Icons.location_city),
-                // 未获得焦点下划线设为灰色
-                enabledBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(color: Colors.grey),
-                ),
-                //获得焦点下划线设为蓝色
-                focusedBorder: UnderlineInputBorder(
-                  borderSide: BorderSide(color: Colors.blue),
-                ),
-              ),
-              onChanged: (value) {
-                _email = value;
-              },
-            ),
-            TextField(
-              decoration: InputDecoration(
-                labelText: "请输入职位描述",
-                prefixIcon: Icon(Icons.details),
+                labelText: "请输入邮箱",
+                prefixIcon: Icon(Icons.mail),
                 // 未获得焦点下划线设为灰色
                 enabledBorder: UnderlineInputBorder(
                   borderSide: BorderSide(color: Colors.grey),
@@ -165,21 +149,21 @@ class _CreatejobshomeState extends State<Createjobshome> {
               child: RaisedButton.icon(
                 icon: Icon(Icons.done_all),
                 label: Text(
-                  "同意招聘协议并发布",
+                  "同意协议并注册",
                   style: TextStyle(
                     fontSize: 10.0,
                   ),
                 ),
                 onPressed: () {
                   _sendReinfo();
-                  runApp(FlMainpage());
+                  runApp(FlLogin());
                 },
               ),
             ),
             Padding(
-              padding: EdgeInsets.only(left: 150, top: 50),
+              padding: EdgeInsets.only(top: 50),
               child: Text(
-                "请仔细阅读《工作协议》和《招聘协议》",
+                "请仔细阅读《用户手册》和《隐私设置》",
                 style: TextStyle(
                   fontSize: 8,
                   color: Colors.indigo,
